Conversion of meteorological data from basin scale to sub-basin scale

Conversion of meteorological data from basin scale to sub-basin scale

Usage

ConvertMeteoSD(x, ...)

# S3 method for class 'GRiwrm'
ConvertMeteoSD(x, meteo, ...)

# S3 method for class 'character'
ConvertMeteoSD(x, griwrm, meteo, ...)

# S3 method for class 'matrix'
ConvertMeteoSD(x, areas, temperature = FALSE, ...)

Arguments

x

either a GRiwrm network description (See CreateGRiwrm), a character id of a node, or a matrix containing meteorological data

...

Parameters passed to the methods

meteo

matrix or data.frame containing meteorological data. Its colnames should be equal to the ID of the basins

griwrm

GRiwrm object describing the semi-distributed network (See CreateGRiwrm)

areas

numeric vector with the total area of the basin followed by the areas of the upstream basins in km2

temperature

logical TRUE if the meteorological data contain air temperature. If FALSE minimum output values are bounded to zero

Value

matrix a matrix containing the converted meteorological data

See also