Creation of the CalibOptions object

This function can be used either for a catchment (with an InputsModel object) or for a network (with a GRiwrmInputsModel object)

# S3 method for GRiwrmInputsModel
CreateCalibOptions(x, FixedParam = NULL, ...)

CreateCalibOptions(x, FixedParam = NULL, ...)

# S3 method for InputsModel
CreateCalibOptions(x, FixedParam = NULL, ...)

# S3 method for character
CreateCalibOptions(x, FixedParam = NULL, ...)

# S3 method for function
CreateCalibOptions(x, FixedParam = NULL, ...)

# S3 method for RunModel_Reservoir
CreateCalibOptions(x, FixedParam = NULL, ...)

Arguments

x

For a single catchment, it can be an object of class InputsModel or a function or a character corresponding to FUN_MOD (compliant with airGR call). For a network, it should be an object of class GRiwrmInputsModel. See CreateInputsModel for details

FixedParam

a numeric vector as for airGR::CreateCalibOptions, or a list giving the values of non-optimised parameters (see details)

...

arguments passed to airGR::CreateCalibOptions, see details

Value

Depending on the class of InputsModel argument (respectively InputsModel and GRiwrmInputsModel object), the returned value is respectively:

a CalibOptions object (See airGR::CreateCalibOptions)

a GRiwrmCalibOptions object which is a list of CalibOptions object with one item per modeled sub-catchment

Details

See airGR::CreateCalibOptions documentation for a complete list of arguments.

With a GRiwrmInputsModel object, all arguments are applied on each sub-catchments of the network with some adaptation depending on the model used on each node.

If the argument FixedParam is a numeric vector, it is applied to each node of the network. Parameters are adapted depending on the use of the routing model and the CemaNeige model on each node. If FixedParam is a list of numeric, each item of the list will be applied on corresponding nodes. Use the id "*" for applying a setting on the remaining nodes. Example for applying one setting for all the nodes except the id "54057":

FixedParam <- list(`*` = c(NA, NA, NA, NA, NA, 0.25, NA, 10, NA),
                   `54057` = c(0.5, NA, NA, NA, NA, 0.25, NA, 10, NA))

The argument IsHyst is ignored since it should be defined previously with CreateInputsModel.GRiwrm.

See also