Extract calibrated parameters

Extract list of parameters from the output of Calibration.GRiwrmInputsModel which can be directly used as argument Param of RunModel.GRiwrmInputsModel, RunModel.Supervisor, and RunModel.GRiwrmOutputsModel.

Usage

extractParam(x)

# S3 method for class 'GRiwrmOutputsCalib'
extractParam(x)

# S3 method for class 'GRiwrmOutputsModel'
extractParam(x)

Arguments

x

Object of class GRiwrmOutputsModel returned by Calibration.GRiwrmInputsModel

Value

A list containing representing the calibrated parameters of each modeled node as a numeric vector.

Details

See vignettes and example of RunModel_Reservoir for examples of use.

See also