Function which creates screen plots giving an overview of the model outputs in the GRiwrm network

Function which creates screen plots giving an overview of the model outputs in the GRiwrm network

# S3 method for class 'GRiwrmOutputsModel'
plot(x, Qobs = NULL, unit = "m3/s", ...)

Arguments

x

[object of class GRiwrmOutputsModel] see RunModel.GRiwrmInputsModel for details

Qobs

(optional) matrix time series of observed flows (for the same time steps than simulated) (mm/time step) with one column by hydrological model output named with the node ID (See CreateGRiwrm for details)

unit

(optional) character flows unit ("m3/s" or "mm")

...

Further arguments for airGR::plot.OutputsModel and plot

Value

list of plots.

Details

For examples of use see topics RunModel.GRiwrmInputsModel, RunModel_Reservoir, and RunModel.Supervisor.