Transfer GR parameters from one donor sub-basin to a receiver sub-basin

This function is used by Calibration.GRiwrmInputsModel for transferring parameters to ungauged nodes and for providing a priori parameters in parameter regularization (See CreateInputsCrit).

Usage

transferGRparams(
  InputsModel,
  Param,
  donor,
  receiver,
  default_param = NULL,
  verbose = FALSE
)

Arguments

InputsModel

Object of class GRiwrmInputsModel, see CreateInputsModel.GRiwrm for details

Param

numeric vector of GR model parameters

donor

character id of the node which gives its parameters

receiver

character id of the node which receives the parameters from the donor

default_param

numeric vector of GR model parameters if parameters are missing from the donor

verbose

logical Add information message on donor and receiver

Value

A numeric vector of transferred parameters

Details

donor and receiver nodes should have the same GR model with the same snow module configuration.

The transfer takes care of:

the presence/absence of hydraulic routing parameters between the donor and the receiver (But default_param should be provided if the receiver has more parameters than the donor)

the transformation of the X4 parameters of GR models. Indeed, this parameter is correlated to the catchment area. Therefore, it has to be rescaled from the surface of the donor catchment to the surface of the receiver catchment, following the relationship suggested by Lobligeois (2014): \(X4_{receiver} = X4_{donor} \times (S_{receiver} / S_{donor})^{0.3}\).

References

Lobligeois, F., 2014. Mieux connaitre la distribution spatiale des pluies améliore-t-il la modélisation des crues ? Diagnostic sur 181 bassins versants français. Thèse de Doctorat, Irstea (Antony), AgroParisTech (Paris), 312 pp. https://webgr.inrae.fr/Media/Files/biblio-theses/2014_lobligeois_these.pdf