Creation of the RunOptions object

This function can be used either for a catchment (with an InputsModel object) or for a network (with a GRiwrmInputsModel object)

# S3 method for class 'GRiwrmInputsModel'
CreateRunOptions(x, IniStates = NULL, ...)

CreateRunOptions(x, ...)

# S3 method for class 'InputsModel'
CreateRunOptions(x, ...)

# S3 method for class 'character'
CreateRunOptions(x, InputsModel, ...)

# S3 method for class 'function'
CreateRunOptions(x, InputsModel, ...)

Arguments

x

For a single catchment, it can be an object of class InputsModel or a function or a character corresponding to FUN_MOD (compliant with airGR call). For a network, it should be an object of class GRiwrmInputsModel. See CreateInputsModel for details

IniStates

(optional) numeric object or list of numeric object of class IniStates, see airGR::CreateIniStates for details

...

arguments passed to airGR::CreateRunOptions, see details

InputsModel

object of class InputsModel (only used to be consistent with the original airGR::CreateRunOptions which has FUN_MOD as first parameter) see airGR::CreateInputsModel for details

Value

Depending on the class of InputsModel argument (respectively InputsModel and GRiwrmInputsModel object), the returned value is respectively:

a RunOptions object (See airGR::CreateRunOptions)

a GRiwrmRunOptions object which is a list of RunOptions objects with one item per modeled sub-catchment

Details

See airGR::CreateRunOptions documentation for a complete list of arguments.

If x argument is a GRiwrmInputsModel object, IniStates must be a list of numeric object of class IniStates with one item per modeled sub-catchment.

With a GRiwrmInputsModel object, all arguments are applied on each sub-catchments of the network.

For examples of use see topics RunModel.GRiwrmInputsModel, RunModel_Reservoir, and RunModel.Supervisor.

See also