Wrapper for airGR::RunModel for one sub-basin

Wrapper for airGR::RunModel for one sub-basin

# S3 method for InputsModel
RunModel(x = NULL, RunOptions, Param, FUN_MOD = NULL, InputsModel = NULL, ...)

Arguments

x

[object of class InputsModel] see airGR::CreateInputsModel for details

RunOptions

[object of class RunOptions] see CreateRunOptions for details

Param

[numeric] vector of model parameters (See details for SD lag model)

FUN_MOD

[function] hydrological model function (e.g. RunModel_GR4J, RunModel_CemaNeigeGR4J)

InputsModel

[object of class InputsModel] see CreateInputsModel for details

...

Further arguments for compatibility with S3 methods

Value

[list] see RunModel_GR4J or RunModel_CemaNeigeGR4J for details.

If InputsModel parameter has been created for using a semi-distributed (SD) lag model (See CreateInputsModel), the list value contains an extra item named QsimDown which is a numeric series of simulated discharge [mm/time step] related to the run-off contribution of the downstream sub-catchment.

Details

This function calls airGR::RunModel (See airGR::RunModel for further details).

The list produced by the function (See Value section of airGR::RunModel_GR4J) is here completed by:

an item $Qsim_m3 storing the simulated discharge series in m3/s

an item $Qover_m3 storing the volumes of over abstraction which occurs when RunModel_Lag warns for negative simulated flows. The latter reflects the volume that was planned to be drawn from the sub-basin but could not be drawn because of the lack of water.