Plot of a Qm3s object (time series of simulated flows)

This function plot time series of flow rate in m3/s. It's a method for object of class "Qm3s" which can be directly called by plot. It can also be called as a function plot.Qm3s if the first parameter has the good format.

# S3 method for class 'Qm3s'
plot(
  x,
  type = "l",
  xlab = "Date",
  ylab = expression("Flow rate (m"^"3" * "/s)"),
  main = "Simulated flows",
  col = grDevices::hcl.colors(ncol(x) - 1, "Zissou 1"),
  legend = colnames(x)[-1],
  legend.cex = 0.7,
  legend.x = "topright",
  legend.y = NULL,
  lty = 1,
  mgp = c(2.5, 1, 0),
  ...
)

Arguments

x

data.frame with a first column with POSIXt dates and followings columns with flows at each node of the network

type

character plot type (See plot.default), default "l"

xlab

character label for the x axis, default to "Date"

ylab

character label for the y axis, default to "Flow (m3/s)"

main

character main title for the plot, default to "Simulated flows"

col

character plotting colors (See par)

legend

character see parameter legend of legend. Set it to NULL to hide the legend

legend.cex

character cex parameter for the text of the legend (See par)

legend.x, legend.y

Legend position, see x and y parameters in graphics::legend

lty

character or numeric The line type (See par)

mgp

The margin line for the axis title, axis labels and axis line (See par)

...

Further arguments to pass to the matplot functions

Value

Screen plot window.

Details

For examples of use see topics RunModel.GRiwrmInputsModel, RunModel_Reservoir, and RunModel.Supervisor.